EUREKA! Websocket subscriptions are working perfectly.
Upon startup, the server subscribes to all users with addresses and then whenever
it receives a transaction, it updates the balance in the database.

This is the stable base to build on.

I think the next step is to build in the qr code generation, that will give
us the pieces for the widget. Then we switch over to bootstrap, etc